+++ spec/telegram/bot/types/base_spec.rb
@@ -18,4 +18,28 @@ RSpec.describe Telegram::Bot::Types::Base do
       expect(subject[text: 'test', number: 1]).to eq(text: 'test', number: 1)
     end
   end
+
+  describe '#[]' do
+    it 'provides access by symbol and string' do
+      object = klass.new(number: 123)
+      expect(object[:number]).to eq(123)
+      expect(object['number']).to eq(123)
+
+      object = klass.new('number' => 456)
+      expect(object[:number]).to eq(456)
+      expect(object['number']).to eq(456)
+    end
+
+    it 'returns nil for missing values' do
+      object = klass.new(number: 123)
+      expect(object[:text]).to eq(nil)
+      expect(object['text']).to eq(nil)
+    end
+
+    it 'raises for invalid field' do
+      object = klass.new(number: 123)
+      expect { object[:number_2] }.to raise_error(klass::MissingAttributeError)
+      expect(object['text']).to eq(nil)
+    end
+  end
 end